Started a Demo for SEMA
I basically started a demo for schroedinger eigenmaps and semi
supervised manifold alignment on a toy example. I was plotting some
data points that appeared to be in separate clusters and I was
basically going through the algorithm and showing how adding the
potential matrix can force some points to be together in proximity in
the new embedding. Still need to do some work on it. But it’s a start.